refactor: shift interface to model folder
