+++ app/models/navLink.ts
@@ -1,0 +1,4 @@
+export interface navLink {
+  label: string;
+  path: string;
+}
