+++ app/models/url.ts
@@ -1,0 +1,5 @@
+export interface Url {
+  label: string;
+  isPDF: boolean;
+  path: string;
+}
